requirements.txt
fastapi
uvicorn
openai==1.98.0
openai-agents==0.1.0
notion-client
python-dotenv
httpx
sqlalchemy
streamlit
requests
uv
notion-client
pydantic 
pydantic-settings
pydantic[email]
mysql-connector-python
strawberry-graphql[fastapi]
python-multipart
passlib[bcrypt]
python-jose[cryptography]
pydub
ffmpeg
phonenumbers
asyncmy